+++ backend/garpix_user/serializers/auth_token_serializer.py
@@ -3,6 +3,8 @@ from django.utils.translation import ugettext_lazy as _
 from rest_framework import serializers
 from rest_framework.authentication import authenticate
 
+from garpix_user.models import UserSession
+
 
 class AuthTokenSerializer(serializers.Serializer):
     username = serializers.CharField(label=_("Username"))
@@ -18,9 +20,19 @@ class AuthTokenSerializer(serializers.Serializer):
         username = username.lower()
 
         if username and password:
-            user = authenticate(request=self.context.get('request'),
+            request = self.context.get('request')
+            user = authenticate(request=request,
                                 username=username, password=password)
-
+            current_user_session = UserSession.get_from_request(request)
+            print('current_user_session', current_user_session)
+            if current_user_session:
+                if (user_user_session := UserSession.objects.filter(
+                        user=user).first()) and user_user_session != current_user_session:
+                    current_user_session.delete()
+                else:
+                    current_user_session.user = user
+                    current_user_session.recognized = UserSession.UserState.REGISTERED
+                    current_user_session.save()
             # The authenticate call simply returns None for is_active=False
             # users. (Assuming the default ModelBackend authentication
             # backend.)