+++ backend/garpix_user/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.2.0 (06.02.2023)
+
+- Russian localization updated
+- Restore password bugs fixed
+- UserSession bugs fixed
+
 ### 3.1.0 (18.01.2023)
 
 - Russian localization added